Expanded view classes into their own modules for easier reading
