+++ cogs/config.py
@@ -1,145 +1,6 @@
 from asyncio import sleep
-import discord.ui as ui
 from lib.checks import *
-
-# Desplegable de selección de rol
-class SelectRole(ui.RoleSelect):
-    def __init__(self, label, maxvals):
-        super().__init__(placeholder=label, max_values=maxvals, min_values=1)
-        self.default_values = []
-
-    async def callback(self, interaction: discord.Interaction):
-        await interaction.response.defer(ephemeral=True)
-
-# Desplegable de selección de canal
-class SelectChannel(ui.ChannelSelect):
-    def __init__(self, label, maxvals):
-        super().__init__(placeholder=label, max_values=maxvals, min_values=1)
-
-    async def callback(self, interaction: discord.Interaction):
-        await interaction.response.defer(ephemeral=True)
-
-# Modal de selección de salario
-class SalarioModal(ui.Modal, title="Salario"):
-    def __init__(self, role_name):
-        super().__init__()
-        self.salario.label = f"Salario de {role_name}:"
-        self.abort = False
-
-    salario = ui.TextInput(label="WIP", style=discord.TextStyle.short, placeholder="1", default="1")
-
-    async def on_submit(self, interaction: discord.Interaction):
-        await interaction.response.defer(ephemeral=True)
-
-    async def on_error(self, interaction: discord.Interaction, error):
-        self.abort = True
-        await interaction.response.defer(ephemeral=True)
-
-# Objeto genérico con botones de continuar y de cancelar
-class ViewObj(ui.View):
-    def __init__(self):
-        super().__init__()
-        self.abort = False
-
-    async def disable(self, interaction: discord.Interaction):
-        for child in self.children:
-            child.disabled = True
-        await interaction.message.edit(content=interaction.message.content, view=self)
-
-    @ui.button(label="Cancelar", style=discord.ButtonStyle.red, row=2)
-    async def cancel(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await interaction.response.send_message("Cancelando configuración...", ephemeral=True, delete_after=2)
-        self.abort = True
-        await self.disable(interaction)
-        self.stop()
-
-    # Función auxiliar presente para overriding
-    async def _cont(self, interaction: discord.Interaction):
-        await interaction.response.defer(ephemeral=True)
-        await self.disable(interaction)
-        self.stop()
-
-    @ui.button(label="Continuar", style=discord.ButtonStyle.green, row=2)
-    async def cont(self, interaction: discord.Interaction, button: discord.ui.Button):
-        await self._cont(interaction)
-
-# Menú de selección de rol
-class SelectRoleView(ViewObj):
-    def __init__(self, label, maxvals=1):
-        super().__init__()
-        self.label = label
-        self.selection = SelectRole(label, maxvals)
-        self.add_item(self.selection)
-
-    async def _cont(self, interaction: discord.Interaction):
-        if not self.selection.values:
-            await interaction.response.send_message(InvalidSelection().response(), ephemeral=True)
-            return
-        await super()._cont(interaction)
-
-# Menú de selección de canal
-class SelectChannelView(ViewObj):
-    def __init__(self, label, maxvals=1):
-        super().__init__()
-        self.label = label
-        self.selection = SelectChannel(label, maxvals)
-        self.add_item(self.selection)
-
-    async def _cont(self, interaction: discord.Interaction):
-        if not self.selection.values:
-            await interaction.response.send_message(InvalidSelection().response(), ephemeral=True)
-            return
-        await super()._cont(interaction)
-
-# Menú de configuración de salarios
-class SalarioView(ViewObj):
-    # Función auxiliar que inicializa el campo de selección de rol
-    def _init_field(self):
-        if hasattr(self, "rol"): self.remove_item(self.rol)
-        self.rol = SelectRole("Selecciona un rol", maxvals=1)
-        self.add_item(self.rol)
-
-    def __init__(self, label):
-        super().__init__()
-        self._init_field()
-        self.d = {}
-        self.label = label
-
-    # Función que añade un diccionario {id_rol: salario} al diccionario del menú y devuelve la selección
-    async def _add_to_d(self, interaction: discord.Interaction):
-        if not self.rol.values: raise InvalidSelection
-        rol = self.rol.values[0]
-        id = rol.id
-        modal = SalarioModal(rol.name)
-
-        await interaction.response.send_modal(modal)
-        await modal.wait()
-
-        d = {id: int(modal.salario.value)}
-        self._init_field() # Reinicia el campo de selección
-        self.d.update(d)
-        return rol, modal.salario.value # Devuelve el rol y el salario para estilizar el mensaje
-
-    # Botón que implementa la función anterior y actualiza el mensaje
-    @ui.button(label="Añadir salario", style=discord.ButtonStyle.blurple, row=2)
-    async def add_to_d(self, interaction: discord.Interaction, button: ui.Button):
-        try: rol, n = await self._add_to_d(interaction)
-        except InvalidSelection as e:
-            await interaction.response.send_message(e.response(), ephemeral=True)
-            return
-        x = lambda i: f"{rol.mention}: {i} Guydocup{"ón" if i==1 else "ones"}."
-        msg = interaction.message.content + "\n" + x(n)
-        await interaction.message.edit(content=msg, view=self)
-
-    # Sobrescritura de la función del botón continuar para obligar a escoger al menos un rol
-    async def _cont(self, interaction):
-        if self.d == {}:
-            try: await self._add_to_d(interaction)
-            except InvalidSelection as e:
-                await interaction.response.send_message(e.response(), ephemeral=True)
-                return
-            await interaction.response.defer(ephemeral=True)
-        await super()._cont(interaction)
+import views.config
 
 
 
@@ -175,10 +36,10 @@ class Config(commands.Cog):
     @commands.hybrid_command(name="configurar", description="Configura los datos del servidor")
     @owner_only()
     async def config(self, ctx: commands.Context):
-        configs = [SelectRoleView("Selecciona los roles de administrador", 25),
-                   SelectChannelView("Selecciona el canal de tienda"), SelectChannelView("Selecciona el canal bancario"),
-                   SalarioView("Selecciona los salarios de los miembros"), SelectRoleView("Selecciona el rol de senador honorario"),
-                   SelectRoleView("Selecciona el rol de esclavo"), SelectRoleView("Selecciona el rol de artista")]
+        configs = [views.config.SelectRoleView("Selecciona los roles de administrador", 25),
+                   views.config.SelectChannelView("Selecciona el canal de tienda"), views.config.SelectChannelView("Selecciona el canal bancario"),
+                   views.config.SalarioView("Selecciona los salarios de los miembros"), views.config.SelectRoleView("Selecciona el rol de senador honorario"),
+                   views.config.SelectRoleView("Selecciona el rol de esclavo"), views.config.SelectRoleView("Selecciona el rol de artista")]
         message = await ctx.reply("Iniciando configuración...")
         await sleep(0.5)
 
@@ -200,7 +61,7 @@ class Config(commands.Cog):
         server.rol_esclavo = configs[5].selection.values[0].id
         server.rol_artista = configs[6].selection.values[0].id
 
-        confirm = ViewObj()
+        confirm = views.config.ViewObj()
         # Embed con los cambios realizados
         embed = self._list_config(server, ctx)
         await message.edit(content="¿Continuar?", embed=embed, view=confirm)
